♻️ [Refactoring]: #041 core/src/document/page-tree/ 配下を 1機能1フォルダ化
- attr-resolver, dict-reader, inheritance-resolver, page-tree-walker, resolved-page を {feature}/index.ts へ
- 各テストを {feature}/__tests__/ へ git mv
- 主実装内の相対 import を +1、test 内を +2 階層調整
- page-tree-walker.test.helpers の cross-feature 参照を `../../page-tree-walker/__tests__/...` に書き換え
- page-tree/index.ts (barrel) は folder/index 解決で無修正
- Test Files=124 / Tests=1688 維持
