+++ packages/core/src/document/page-tree/page-tree-walker/__tests__/page-tree-walker.test.helpers.ts
@@ -1,18 +1,18 @@
 import { expect, type Mock, vi } from "vitest";
-import type { PdfError } from "../../pdf/errors/error/index";
-import { GenerationNumber } from "../../pdf/types/generation-number/index";
-import { ObjectNumber } from "../../pdf/types/object-number/index";
+import type { PdfError } from "../../../../pdf/errors/error/index";
+import { GenerationNumber } from "../../../../pdf/types/generation-number/index";
+import { ObjectNumber } from "../../../../pdf/types/object-number/index";
 import type {
   IndirectRef,
   PdfDictionary,
   PdfIndirectRef,
   PdfObject,
   PdfValue,
-} from "../../pdf/types/pdf-types/index";
-import type { Result } from "../../utils/result/index";
-import { err, ok } from "../../utils/result/index";
-import type { ResolveRef } from "../catalog-parser";
-import type { PdfRectangle } from "./resolved-page";
+} from "../../../../pdf/types/pdf-types/index";
+import type { Result } from "../../../../utils/result/index";
+import { err, ok } from "../../../../utils/result/index";
+import type { ResolveRef } from "../../../catalog-parser";
+import type { PdfRectangle } from "../../resolved-page";
 
 /**
  * Result が Ok であることを `expect` で保証し、値を返す（テスト専用ヘルパ）。
